TravelChatbot.App/requirements.txt
openai
streamlit
tenacity
python-dotenv
boto3
botocore
chromadb
numpy
pandas